Fixed IOU Score Calculations & Added Object Labels
Fixed IOU score calculations which were occuring since the example boxes were not in the image coordinate axes. The y-axis was flipped, also added batch handling to the function since the data being passed will be done in batches. Also created a method to add labels for foreground, background. Added a '-1' label to allow for filtering easily to reduce the noise in the training data. Need to extensively test out the label assignment and fix the overlap utility function to handle batches.